Remove double buffering
This is the only sane design for async I/O on a webserver modulo a few minor
tweaks. Everything else leads us down the rabbit hole of trying to implement
page cache in userspace and doing a bad job. We're still going to incur the
hard page faults and, worse, end up putting fs data in to swap.

As it happens, AIO gives nothing but a slight drop in throughput and increase
in latency when the working set fits in page cache. All this at the cost of
massively increased system CPU time.

Oh well :)