Slowly, daylight crept in through the bandages
Earlier problems were a result of double-free on the IOCB combined with a
possibly bogus rbtree delete implementation. If rbtree_delete() is ever called,
IOW if cache size < webroot then things go tits up very quickly.

Got rid of a lot of the paranoia but left in all the legit assertions.

Also use in_flight to check for spurious eventfd wake-ups. Pretty sure this is
a legit/expected case of racing between the eventfd, kaio and io_getevents()
and it saves us an unnecessary system call.