Re-work async state-machine
Firstly, put calling context to sleep in io_submit rather than waiting for
a write event to decide that we don't want to handle a write event.

Secondly avoid frobbing the h_data_off and h_data_len fields in the completion
callback context. This creates bogosity for subsequent io_submit(), or was it
in HTTP state machine? Can't remember...