Eliminate redundant epoll_ctl() syscalls
There are two cases were extraneous epoll_ctl() syscalls were being
generated. Firstly EPOLL_CTL_DEL on a file descriptor that is marked for death.
Solve this by calling close() on it before nbio_del() and setting it to -1.
Epoll plugin then checks for this condition before removing it from the fd set.

Secondly, when pump() is called, we know the active queue is empty so no need
to EPOLL_CTL_DEL anything. Instead modify nbio_pump() to iterative the inactive
list and call plugin->inactive() on each item before calling plugin->pump().
This gives epoll plugin chance to add to the fd-set if necessary.

Also added EEK comment to poll plugin for the "out of memory on fd set
allocation" issue.