Add http_listener structure
Keeps track of webroots on a per-listener basis. Modified inet_listener() to
return a pointer to the listener object. The web listener keeps a pointer to
the default webroot. All http-conns keep a pointer to the web listener object.

Eventually each listener will keep a map of all the vhosts and such and
point to different webroot objects.