critbit: Change API to allow handling of value pointers
When inserting an item take a triple-indirect pointer which stores
a pointer to the value-pointer. The caller can check for NULL to
see if an item was already allocated there and decide what to
assign to it.

For delete, take a double-indirect pointer which returns the
value of the deleted item.

For search, take a double-indirect pointer which holds the pointer
to the found nodes value so that it can be examined or updated.

For delete, take a callback function which gets passed the
value pointers of all the deleted nodes.

All of the value manipulation stuff is optional, pass NULL
if you don't care about the map values.