First cut of directory indexing
We create a tmp file to store content for the indexes and just append
each index page to the end of that. The tmpfile is pre-deleted to ensure
cleanup if the mkroot process is terminated unexpectedly.

Then we just indicate in the file records whether we are to read a
file or read a section out of the tmpfile.

The path_splice() code was modified to prevent extraneous slashes
when we construct the links in the indexes. Seems like a good idea
anyway.

Also, write_files() was modified to fix the condition where we end
up with a bogus webroot in some instancess if files were modified
during the process. However it still breaks if files are truncated
downwards so we probably ought to detect that condition later.